+++ source/include/Common/MaaTypes.h
@@ -161,7 +161,7 @@ public:
 
     virtual void set_anchor(const std::string& anchor_name, const std::string& node_name) = 0;
     virtual std::optional<std::string> get_anchor(const std::string& anchor_name) const = 0;
-    virtual uint get_hit_count(const std::string& node_name) const = 0;
+    virtual size_t get_hit_count(const std::string& node_name) const = 0;
     virtual void clear_hit_count(const std::string& node_name) = 0;
 };
